refactor: xmodule/editing_module.py -> xmodule/editing_block.py
