+++ xmodule/annotatable_block.py
@@ -10,7 +10,7 @@ from xblock.core import XBlock
 from xblock.fields import Scope, String
 
 from openedx.core.djangolib.markup import HTML, Text
-from xmodule.editing_module import EditingMixin
+from xmodule.editing_block import EditingMixin
 from xmodule.raw_module import RawMixin
 from xmodule.util.xmodule_django import add_webpack_to_fragment
 from xmodule.xml_module import XmlMixin
