+++ xmodule/template_module.py
@@ -8,7 +8,7 @@ from xblock.core import XBlock
 from lxml import etree
 from pkg_resources import resource_string
 from web_fragments.fragment import Fragment
-from xmodule.editing_module import EditingMixin
+from xmodule.editing_block import EditingMixin
 from xmodule.raw_module import RawMixin
 from xmodule.util.xmodule_django import add_webpack_to_fragment
 from xmodule.x_module import (
